LPS-179765 Extract LayoutCPDefinitionDisplayLayoutSelectionScreenNavigationCategory methods to a LayoutCPDefinitionDisplayLayoutSelectionScreenNavigationEntry component and extends LayoutCPDefinitionDisplayLayoutSelectionScreenNavigationCategory in it
